ref: move user classes to models
