+++ src/models/index.ts
@@ -1,1 +1,2 @@
 export * from './message.model';
+export * from './user.models';
