+++ src/models/user.models.ts
@@ -1,0 +1,22 @@
+import { randomUUID } from 'crypto';
+
+export interface IUser {
+  index: string;
+  name: string;
+  password: string;
+}
+
+export class User implements IUser {
+  index: string = '';
+
+  name: string = '';
+
+  password: string = '';
+
+  constructor(user: Partial<User>) {
+    // TODO: validate
+    Object.assign(this, user);
+
+    this.index = randomUUID();
+  }
+}
